+++ drugex/about.py
@@ -5,4 +5,4 @@ Created by: Martin Sicho
 On: 24.06.22, 10:36
 """
 
-VERSION = "3.3.0.dev0"
+VERSION = "3.3.0"