+++ testing/clitest/test_single.sh
@@ -8,7 +8,7 @@ ${DATASET_COMMON_ARGS} \
 -i ${TEST_DATA_PRETRAINING} \
 -o ${PRETRAINING_PREFIX} \
 -mt smiles \
--sm \
+-vf 'voc_smiles.txt' \
 -nof
 echo "Test: Done."
 
@@ -17,8 +17,8 @@ python -m drugex.dataset \
 ${DATASET_COMMON_ARGS} \
 -i ${TEST_DATA_FINETUNING} \
 -o ${FINETUNING_PREFIX} \
--sm \
--mt smi \
+-mt smiles \
+-vf 'voc_smiles.txt' \
 -nof
 echo "Test: Done."