+++ CHANGELOG.md
@@ -16,12 +16,16 @@ None.
   - arguments `drop_duplicates`, `drop_invalid` were added,
   - argument `keep_frags` was added.
 - The `sample_smiles` method was added to the SequenceTranformer `GTP2Model` and to the `RNN` classes.
+- Changed the `GTP2Model` adaptive learning rate settings to resolve pretraining issues
 - Progress bars were added for models' fitting (pretraining, fine-tuning and reinforcement learning).
 - Tokens `_` and `.` always present in `VocSmiles` have been removed.
 - RNN models deposited on Zenodo and pretrained on ChEMBL31 and Papyrus 05.5 were updated while the RNN model pretrained on ChEMBL27 did not need to.
+- Moved encoding of tokens for SMILES-based models to the parallel preprocessing steps to improve performance
+- All testing code that is not unit tests was moved to `testing`
 - Remove QSAR modelling from DrugEx, now in a seperate repository QSPRpred
 
 
 ## New Features
 
 - Tutorial for scaffold-based generation.
+- Added tests to `testing` that allow to check consistency of models between versions.